+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 0.9.50
  * stax rebase - now supports resumable rebase
+ * stax amend - now has working rebase
 
 0.9.49
  * stax help - now shows command short names too